v2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "beyan-agentic"
version = "2.0.0"
description = "A powerful, modular agentic framework for repository analysis and refactoring."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "XINMurat", email = "murat@example.com"}
]
keywords = ["agentic", "ai", "refactoring", "analysis", "llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "tiktoken>=0.6.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
beyan = "analyzer:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
